refactored miscobjects, pluginobjects to own files for tidyness
